# ARROW-2653: [C++] Refactor hash table support