--- a/cpp/src/arrow/compute/kernels/hash.h
+++ b/cpp/src/arrow/compute/kernels/hash.h
@@ -38,6 +38,8 @@ class FunctionContext;
 /// values. Implementations should be thread-safe
 class ARROW_EXPORT HashKernel : public UnaryKernel {
  public:
+  // XXX why are those methods exposed?
+  virtual Status Reset() = 0;
   virtual Status Append(FunctionContext* ctx, const ArrayData& input) = 0;
   virtual Status Flush(Datum* out) = 0;
   virtual Status GetDictionary(std::shared_ptr<ArrayData>* out) = 0;