--- a/cpp/src/arrow/util/CMakeLists.txt
+++ b/cpp/src/arrow/util/CMakeLists.txt
@@ -37,7 +37,7 @@ install(FILES
   date.h
   decimal.h
   hash-util.h
-  hash.h
+  hashing.h
   io-util.h
   key_value_metadata.h
   lazy.h
@@ -89,6 +89,7 @@ ADD_ARROW_TEST(bit-util-test)
 ADD_ARROW_TEST(checked-cast-test)
 ADD_ARROW_TEST(compression-test)
 ADD_ARROW_TEST(decimal-test)
+ADD_ARROW_TEST(hashing-test)
 ADD_ARROW_TEST(key-value-metadata-test)
 ADD_ARROW_TEST(lazy-test)
 ADD_ARROW_TEST(logging-test)
@@ -102,6 +103,7 @@ ADD_ARROW_TEST(utf8-util-test)
 ADD_ARROW_BENCHMARK(bit-util-benchmark)
 ADD_ARROW_BENCHMARK(compression-benchmark)
 ADD_ARROW_BENCHMARK(decimal-benchmark)
+ADD_ARROW_BENCHMARK(hashing-benchmark)
 ADD_ARROW_BENCHMARK(lazy-benchmark)
 ADD_ARROW_BENCHMARK(number-parsing-benchmark)
 ADD_ARROW_BENCHMARK(utf8-util-benchmark)