--- a/dev/release/rat_exclude_files.txt
+++ b/dev/release/rat_exclude_files.txt
@@ -20,6 +20,8 @@ cpp/src/arrow/util/variant/recursive_wrapper.h
 cpp/src/arrow/util/variant/variant_cast.h
 cpp/src/arrow/util/variant/variant_io.h
 cpp/src/arrow/util/variant/variant_visitor.h
+cpp/src/arrow/util/xxhash/xxhash.c
+cpp/src/arrow/util/xxhash/xxhash.h
 cpp/build-support/asan_symbolize.py
 cpp/build-support/cpplint.py
 cpp/build-support/clang_format_exclusions.txt
@@ -41,8 +43,6 @@ cpp/src/plasma/thirdparty/ae/ae_select.c
 cpp/src/plasma/thirdparty/ae/config.h
 cpp/src/plasma/thirdparty/ae/zmalloc.h
 cpp/src/plasma/thirdparty/dlmalloc.c
-cpp/src/plasma/thirdparty/xxhash.cc
-cpp/src/plasma/thirdparty/xxhash.h
 dev/release/rat_exclude_files.txt
 dev/tasks/linux-packages/debian.ubuntu-trusty/compat
 dev/tasks/linux-packages/debian.ubuntu-trusty/control