@pitrou 2018-11-20
Dict building benchmark:
* before:
```$ ./build/release/builder-benchmark --benchmark_filter=Dict
[...]
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                   185 us        185 us       3701   411.539MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                   187 us        187 us       3701   407.759MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean              186 us        186 us       3701   409.649MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median            186 us        186 us       3701   409.649MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              1 us          1 us       3701   2.67294MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               187 us        187 us       3696   408.102MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               183 us        183 us       3696   416.262MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean          185 us        185 us       3696   412.182MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median        185 us        185 us       3696   412.182MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          3 us          3 us       3696   5.77062MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  194 us        194 us       3617   394.101MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  193 us        193 us       3617   394.449MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean             194 us        194 us       3617   394.275MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median           194 us        194 us       3617   394.275MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             0 us          0 us       3617   251.506kB/s
BM_BuildStringDictionaryArray/repeats:2                        524 us        524 us       1342    190.84MB/s
BM_BuildStringDictionaryArray/repeats:2                        525 us        525 us       1342    190.48MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   524 us        524 us       1342    190.66MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 524 us        524 us       1342    190.66MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   1 us          1 us       1342   260.545kB/s
```

* after:
```
-----------------------------------------------------------------------------------------------
Benchmark                                                        Time           CPU Iterations
-----------------------------------------------------------------------------------------------
BM_BuildInt64DictionaryArrayRandom/repeats:2                   168 us        168 us       4164   453.871MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2                   168 us        168 us       4164   453.462MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_mean              168 us        168 us       4164   453.667MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_median            168 us        168 us       4164   453.667MB/s
BM_BuildInt64DictionaryArrayRandom/repeats:2_stddev              0 us          0 us       4164   295.764kB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               159 us        159 us       4392   478.758MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2               160 us        160 us       4392   478.048MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_mean          160 us        159 us       4392   478.403MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_median        160 us        159 us       4392   478.403MB/s
BM_BuildInt64DictionaryArraySequential/repeats:2_stddev          0 us          0 us       4392   514.183kB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  160 us        160 us       4372   476.579MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2                  160 us        160 us       4372   476.578MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_mean             160 us        160 us       4372   476.578MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_median           160 us        160 us       4372   476.578MB/s
BM_BuildInt64DictionaryArraySimilar/repeats:2_stddev             0 us          0 us       4372    797.195B/s
BM_BuildStringDictionaryArray/repeats:2                        383 us        383 us       1835   261.189MB/s
BM_BuildStringDictionaryArray/repeats:2                        383 us        382 us       1835   261.343MB/s
BM_BuildStringDictionaryArray/repeats:2_mean                   383 us        383 us       1835   261.266MB/s
BM_BuildStringDictionaryArray/repeats:2_median                 383 us        383 us       1835   261.266MB/s
BM_BuildStringDictionaryArray/repeats:2_stddev                   0 us          0 us       1835   111.316kB/s
```