@pitrou 2018-11-20
Compute benchmark:
* before:
```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       1570 us       1569 us        898   2.48717GB/s
BM_BuildStringDictionary/min_time:1.000                                 4435 us       4434 us        316   68.0927MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             54623 us      54602 us         26    2.2884GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time          100073 us     100038 us         14   1.24908GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time         136464 us     136424 us         10   937.977MB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       804117 us     803776 us          2   159.181MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           81952 us      81918 us         17   1.52528GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time        134001 us     133959 us         10   955.218MB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time       177680 us     177629 us          8   720.398MB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     972930 us     972477 us          2   131.561MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           209702 us     209629 us          7   762.988MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         266693 us     266613 us          5   599.942MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        362979 us     362893 us          4   440.797MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     2886708 us    2885530 us          1   55.4265MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          737258 us     736968 us          2   2.11934GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        867057 us     866734 us          2   1.80207GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time      1143734 us    1143263 us          1   1.36614GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    6824159 us    6821196 us          1   234.461MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time            11718 us      11713 us        119   1.33345GB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          32380 us      32371 us         44    494.13MB/s
```

* after:
```
---------------------------------------------------------------------------------------------------------
Benchmark                                                                  Time           CPU Iterations
---------------------------------------------------------------------------------------------------------
BM_BuildDictionary/min_time:1.000                                       1352 us       1352 us       1033   2.88675GB/s
BM_BuildStringDictionary/min_time:1.000                                 5185 us       5182 us        270   58.2594MB/s
BM_UniqueInt64NoNulls/16777216/50/min_time:1.000/real_time             30674 us      30593 us         46   4.07515GB/s
BM_UniqueInt64NoNulls/16777216/1024/min_time:1.000/real_time           32040 us      31970 us         44   3.90134GB/s
BM_UniqueInt64NoNulls/16777216/10240/min_time:1.000/real_time          43753 us      43733 us         32   2.85697GB/s
BM_UniqueInt64NoNulls/16777216/1048576/min_time:1.000/real_time       516782 us     516368 us          3   247.686MB/s
BM_UniqueInt64WithNulls/16777216/50/min_time:1.000/real_time           48618 us      48597 us         29   2.57107GB/s
BM_UniqueInt64WithNulls/16777216/1024/min_time:1.000/real_time         48776 us      48754 us         29   2.56271GB/s
BM_UniqueInt64WithNulls/16777216/10240/min_time:1.000/real_time        63645 us      63616 us         22   1.96402GB/s
BM_UniqueInt64WithNulls/16777216/1048576/min_time:1.000/real_time     614967 us     614258 us          2   208.141MB/s
BM_UniqueString10bytes/16777216/50/min_time:1.000/real_time           212362 us     212290 us          7    753.43MB/s
BM_UniqueString10bytes/16777216/1024/min_time:1.000/real_time         220632 us     220554 us          6   725.189MB/s
BM_UniqueString10bytes/16777216/10240/min_time:1.000/real_time        338742 us     338649 us          4   472.335MB/s
BM_UniqueString10bytes/16777216/1048576/min_time:1.000/real_time     2683194 us    2682169 us          1   59.6304MB/s
BM_UniqueString100bytes/16777216/50/min_time:1.000/real_time          786376 us     786071 us          2   1.98696GB/s
BM_UniqueString100bytes/16777216/1024/min_time:1.000/real_time        877040 us     876709 us          2   1.78156GB/s
BM_UniqueString100bytes/16777216/10240/min_time:1.000/real_time      1137608 us    1137168 us          1    1.3735GB/s
BM_UniqueString100bytes/16777216/1048576/min_time:1.000/real_time    6099277 us    6096136 us          1   262.326MB/s
BM_UniqueUInt8NoNulls/16777216/200/min_time:1.000/real_time            13559 us      13555 us        103   1.15233GB/s
BM_UniqueUInt8WithNulls/16777216/200/min_time:1.000/real_time          23543 us      23536 us         59   679.612MB/s
```